# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Slots Sports Ground API",
        "description": "REST API для бронирования временных слотов на спортивной площадке.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Основной сервер"
        }
    ],
    "paths": {
        "/api/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Получить список бронирований текущего пользователя",
                "operationId": "4535e967ef15bc76dadf1efa40192f0d",
                "responses": {
                    "200": {
                        "description": "Список бронирований",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookingDTO"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован"
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Создать бронирование с несколькими слотами",
                "operationId": "43e6dcca842886074084b8db86af7802",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookingDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Бронирование создано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingDTO"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Конфликт слотов"
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        },
        "/api/bookings/{booking}/slots/{slot}": {
            "patch": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Обновить слот бронирования",
                "operationId": "6fedabb639be7158e11d7cff12d9d074",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "slot",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookingSlotDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Слот обновлён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingSlotDTO"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён"
                    },
                    "422": {
                        "description": "Конфликт слотов"
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        },
        "/api/bookings/{booking}/slots": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Добавить слот в бронирование",
                "operationId": "f87cbf3c923e444582242e05240261d2",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookingSlotDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Слот добавлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingSlotDTO"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Конфликт слотов"
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        },
        "/api/bookings/{booking}": {
            "delete": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Удалить бронирование",
                "operationId": "dd0cd74ecc75a87121ba3e03ee226677",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Бронирование удалено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Booking deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён"
                    },
                    "404": {
                        "description": "Бронирование не найдено"
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "BookingDTO": {
                "required": [
                    "slots"
                ],
                "properties": {
                    "slots": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BookingSlotDTO"
                        }
                    }
                },
                "type": "object"
            },
            "BookingSlotDTO": {
                "required": [
                    "start_time",
                    "end_time"
                ],
                "properties": {
                    "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-25T12:00:00"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-25T13:00:00"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "api_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Bookings",
            "description": "Операции с бронированиями"
        }
    ]
}